Day 18, corrected algorithm
Adding positions as seen when pushing them on the priority queue is,
incorrect, even though it worked for my input.

Now the possibility of shorter path to the same position is taken into
account by checking against the best distance so far.

The seen set could have been updated whenever an item is popped off the
queue, but that is much slower. I suppose the speedup is from doing less
insertions into the queue, which also keeps the queue shorter and thus
the operations on it faster. Memory overhead should also be reduced by
this.

The approach taken kept the queue at under 400 entries for the first
part and under 120 000 for the second. With the alternative solution it
got to 300 000 in the first part before I stopped it.